fix: let everyone see apple pay cta (#3616)
